refactor: Move display_jobs function to job module and update imports in main.rs
