+++ src/main.rs
@@ -8,7 +8,7 @@ mod sysexits;
 use crate::application::{Application, init_config};
 use crate::file_util::*;
 use crate::item::{get_item, get_items};
-use crate::job::{BackupModel, CompressFormat, Job, Level};
+use crate::job::{BackupModel, CompressFormat, Job, Level, display_jobs};
 use anyhow::{Result, anyhow};
 use clap::{Parser, Subcommand, ValueEnum};
 use futures::StreamExt;
@@ -451,60 +451,6 @@ fn list_by_lte(id: u32) {
     println!("{}", display_jobs(jobs));
 }
 
-fn display_jobs(jobs: Vec<Job>) -> String {
-    if jobs.is_empty() {
-        return String::new();
-    }
-    let mut s = String::from('[');
-    for job in jobs {
-        let comp = match job.compression {
-            Some(CompressFormat::Gzip) => "Gzip",
-            Some(CompressFormat::Zip) => "Zip",
-            Some(CompressFormat::Sevenz) => "Sevenz",
-            Some(CompressFormat::Zstd) => "Zstd",
-            Some(CompressFormat::Bzip2) => "Bzip2",
-            Some(CompressFormat::Xz) => "Xz",
-            Some(CompressFormat::Lz4) => "Lz4",
-            Some(CompressFormat::Tar) => "Tar",
-            None => "",
-        };
-        let level = match job.level {
-            Some(Level::Fastest) => "Fastest",
-            Some(Level::Faster) => "Faster",
-            Some(Level::Default) => "Default",
-            Some(Level::Better) => "Better",
-            Some(Level::Best) => "Best",
-            None => "",
-        };
-        let model = match job.model {
-            Some(BackupModel::Full) => "Full",
-            Some(BackupModel::Mirror) => "Mirror",
-            None => "",
-        };
-        s.push_str(&format!(
-            "{{\n    id: {},\n    source: \"{}\",\n    target: \"{}\"",
-            job.id,
-            job.source.display(),
-            job.target.display()
-        ));
-        if !comp.is_empty() {
-            s.push_str(&format!(",\n    compression: \"{comp}\""));
-        }
-        if !level.is_empty() {
-            s.push_str(&format!(",\n    level: \"{level}\""));
-        }
-        if let Some(ignore) = &job.ignore {
-            s.push_str(&format!(",\n    ignore: {ignore:?}"));
-        }
-        if !model.is_empty() {
-            s.push_str(&format!(",\n    model: \"{model}\""));
-        }
-        s.push_str("\n}");
-    }
-    s.push(']');
-    s
-}
-
 /// Deletes a job by id or deletes all jobs.
 fn delete(id: Option<Vec<u32>>, all: bool) -> Result<()> {
     if all {
@@ -700,45 +646,5 @@ fn canonicalize(path: PathBuf) -> PathBuf {
             }
             process::exit(1);
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::display_jobs;
-    use crate::job::{CompressFormat, Job, Level};
-    use std::path::PathBuf;
-
-    #[test]
-    fn test_job_list_display() {
-        let jobs = vec![
-            Job {
-                id: 1,
-                source: PathBuf::from("/test/source1"),
-                target: PathBuf::from("/test/target1"),
-                compression: Some(CompressFormat::Zip),
-                level: Some(Level::Fastest),
-                ignore: None,
-                model: None,
-            },
-            Job {
-                id: 2,
-                source: PathBuf::from("/test/source2"),
-                target: PathBuf::from("/test/target2"),
-                compression: Some(CompressFormat::Zstd),
-                level: Some(Level::Best),
-                ignore: Some(vec!["*.tmp".to_string()]),
-                model: None,
-            },
-        ];
-
-        let display_str = display_jobs(jobs);
-
-        assert!(display_str.starts_with('['));
-        assert!(display_str.ends_with(']'));
-        assert!(display_str.contains("id: 1"));
-        assert!(display_str.contains("id: 2"));
-        assert!(display_str.contains("Zip"));
-        assert!(display_str.contains("Zstd"));
     }
 }
